Typography
On Unfinished Business, Andy mentioned wanting to improve the
typography scss. I was trying to make it into one function but couldn't
work out the maths that you used to work out each heading.

I have however created a mixin that makes it a little neater… but I
think we can do better!